requirements.txt
pytest==6.2.2
allure-pytest==2.8.34